Update ValuesJsonConverter for System.Text.Json
This is bound to have issues as a number of subtle changes have occurred. Because the value is unable to be determined, we use the token type instead for "ParseTokenValueArguments".

Moderate sized sections of DateTime handling code had to be commented out as there doesn't seem to be a similar path.