Updated schema tool building logic
This specifically drops Newtonsoft from the string-building process for the schema classes.

Of note with this change is the larger block of code in Class.cs with adding the `Context` property on `Thing`. This change is necessary as there is a collision between this specifically added property and the `Context` property on `JsonLdObject`.