Updated handling of additional context removal
With the root context no longer being at the start of the JSON, it needs to be handled differently. Because properties are now tied to inheritance and `Thing` has the `@Context` property, we can instead use `LastIndexOf` as the way to determine where the root context information is. We use this value as the length from the start to narrow the search-and-replace space.